LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/text-plain.window-expected.txt


FAIL text/plain: Basic test (normal form) assert_equals: expected "basic=test\r\n" but got "basic=test"
FAIL text/plain: Basic test (formdata event) assert_equals: expected "basic=test\r\n" but got ""
FAIL text/plain: Basic File test (normal form) assert_equals: expected "basic=file-test.txt\r\n" but got "basic=file-test.txt"
FAIL text/plain: Basic File test (formdata event) assert_equals: expected "basic=file-test.txt\r\n" but got ""
FAIL text/plain: 0x00 in name (normal form) assert_equals: expected "a\0b=c\r\n" but got "a\0b=c"
FAIL text/plain: 0x00 in name (formdata event) assert_equals: expected "a\0b=c\r\n" but got ""
FAIL text/plain: 0x00 in value (normal form) assert_equals: expected "a=b\0c\r\n" but got "a=b\0c"
FAIL text/plain: 0x00 in value (formdata event) assert_equals: expected "a=b\0c\r\n" but got ""
FAIL text/plain: 0x00 in filename (normal form) assert_equals: expected "a=b\0c\r\n" but got "a=b\0c"
FAIL text/plain: 0x00 in filename (formdata event) assert_equals: expected "a=b\0c\r\n" but got ""
FAIL text/plain: \n in name (normal form) assert_equals: expected "a\r\nb=c\r\n" but got "a\r\nb=c"
FAIL text/plain: \n in name (formdata event) assert_equals: expected "a\r\nb=c\r\n" but got ""
FAIL text/plain: \r in name (normal form) assert_equals: expected "a\r\nb=c\r\n" but got "a\r\nb=c"
FAIL text/plain: \r in name (formdata event) assert_equals: expected "a\r\nb=c\r\n" but got ""
FAIL text/plain: \r\n in name (normal form) assert_equals: expected "a\r\nb=c\r\n" but got "a\r\nb=c"
FAIL text/plain: \r\n in name (formdata event) assert_equals: expected "a\r\nb=c\r\n" but got ""
FAIL text/plain: \n\r in name (normal form) assert_equals: expected "a\r\n\r\nb=c\r\n" but got "a\r\n\r\nb=c"
FAIL text/plain: \n\r in name (formdata event) assert_equals: expected "a\r\n\r\nb=c\r\n" but got ""
FAIL text/plain: \n in value (normal form) assert_equals: expected "a=b\r\nc\r\n" but got "a=b\r\nc"
FAIL text/plain: \n in value (formdata event) assert_equals: expected "a=b\r\nc\r\n" but got ""
FAIL text/plain: \r in value (normal form) assert_equals: expected "a=b\r\nc\r\n" but got "a=b\r\nc"
FAIL text/plain: \r in value (formdata event) assert_equals: expected "a=b\r\nc\r\n" but got ""
FAIL text/plain: \r\n in value (normal form) assert_equals: expected "a=b\r\nc\r\n" but got "a=b\r\nc"
FAIL text/plain: \r\n in value (formdata event) assert_equals: expected "a=b\r\nc\r\n" but got ""
FAIL text/plain: \n\r in value (normal form) assert_equals: expected "a=b\r\n\r\nc\r\n" but got "a=b\r\n\r\nc"
FAIL text/plain: \n\r in value (formdata event) assert_equals: expected "a=b\r\n\r\nc\r\n" but got ""
FAIL text/plain: \n in filename (normal form) assert_equals: expected "a=b\r\nc\r\n" but got "a=b\r\nc"
FAIL text/plain: \n in filename (formdata event) assert_equals: expected "a=b\r\nc\r\n" but got ""
FAIL text/plain: \r in filename (normal form) assert_equals: expected "a=b\r\nc\r\n" but got "a=b\r\nc"
FAIL text/plain: \r in filename (formdata event) assert_equals: expected "a=b\r\nc\r\n" but got ""
FAIL text/plain: \r\n in filename (normal form) assert_equals: expected "a=b\r\nc\r\n" but got "a=b\r\nc"
FAIL text/plain: \r\n in filename (formdata event) assert_equals: expected "a=b\r\nc\r\n" but got ""
FAIL text/plain: \n\r in filename (normal form) assert_equals: expected "a=b\r\n\r\nc\r\n" but got "a=b\r\n\r\nc"
FAIL text/plain: \n\r in filename (formdata event) assert_equals: expected "a=b\r\n\r\nc\r\n" but got ""
FAIL text/plain: double quote in name (normal form) assert_equals: expected "a\"b=c\r\n" but got "a\"b=c"
FAIL text/plain: double quote in name (formdata event) assert_equals: expected "a\"b=c\r\n" but got ""
FAIL text/plain: double quote in value (normal form) assert_equals: expected "a=b\"c\r\n" but got "a=b\"c"
FAIL text/plain: double quote in value (formdata event) assert_equals: expected "a=b\"c\r\n" but got ""
FAIL text/plain: double quote in filename (normal form) assert_equals: expected "a=b\"c\r\n" but got "a=b\"c"
FAIL text/plain: double quote in filename (formdata event) assert_equals: expected "a=b\"c\r\n" but got ""
FAIL text/plain: single quote in name (normal form) assert_equals: expected "a'b=c\r\n" but got "a'b=c"
FAIL text/plain: single quote in name (formdata event) assert_equals: expected "a'b=c\r\n" but got ""
FAIL text/plain: single quote in value (normal form) assert_equals: expected "a=b'c\r\n" but got "a=b'c"
FAIL text/plain: single quote in value (formdata event) assert_equals: expected "a=b'c\r\n" but got ""
FAIL text/plain: single quote in filename (normal form) assert_equals: expected "a=b'c\r\n" but got "a=b'c"
FAIL text/plain: single quote in filename (formdata event) assert_equals: expected "a=b'c\r\n" but got ""
FAIL text/plain: backslash in name (normal form) assert_equals: expected "a\\b=c\r\n" but got "a\\b=c"
FAIL text/plain: backslash in name (formdata event) assert_equals: expected "a\\b=c\r\n" but got ""
FAIL text/plain: backslash in value (normal form) assert_equals: expected "a=b\\c\r\n" but got "a=b\\c"
FAIL text/plain: backslash in value (formdata event) assert_equals: expected "a=b\\c\r\n" but got ""
FAIL text/plain: backslash in filename (normal form) assert_equals: expected "a=b\\c\r\n" but got "a=b\\c"
FAIL text/plain: backslash in filename (formdata event) assert_equals: expected "a=b\\c\r\n" but got ""
FAIL text/plain: non-ASCII in name and value (normal form) assert_equals: expected "Ã¡b=Ã§\r\n" but got "Ã¡b=Ã§"
FAIL text/plain: non-ASCII in name and value (formdata event) assert_equals: expected "Ã¡b=Ã§\r\n" but got ""
FAIL text/plain: non-ASCII in filename (normal form) assert_equals: expected "a=É.txt\r\n" but got "a=É.txt"
FAIL text/plain: non-ASCII in filename (formdata event) assert_equals: expected "a=É.txt\r\n" but got ""
FAIL text/plain: characters not in encoding in name and value (normal form) assert_equals: expected "a&#601;b=c&#65533;d\r\n" but got "a&#601;b=c&#65533;d"
FAIL text/plain: characters not in encoding in name and value (formdata event) assert_equals: expected "a&#601;b=c&#65533;d\r\n" but got ""
FAIL text/plain: character not in encoding in filename (normal form) assert_equals: expected "á=&#128169;\r\n" but got "á=&#128169;"
FAIL text/plain: character not in encoding in filename (formdata event) assert_equals: expected "á=&#128169;\r\n" but got ""
